renamed sidebar to `LeadingSidebar`
