+++ CodeEdit/LeadingSidebar/Navigator/NavigatorSidebarItem.swift
