+++ CodeEdit/LeadingSidebar/Search/SearchBar.swift
