+++ CodeEdit/LeadingSidebar/Search/SearchModeSelector.swift
