+++ CodeEdit/LeadingSidebar/Search/SearchResultFileItem.swift
