+++ CodeEdit/LeadingSidebar/Search/SearchResultList.swift
