+++ CodeEdit/LeadingSidebar/Search/SidebarSearch.swift
